LayoutTests/svg/dom/rgb-color-parser-expected.txt

This test fuzzes the length list parser with semi-random attribute values and dumps the results of any values that parse successfully.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(.78-2e8e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(% 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(7
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(e+
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb( -2e9%0.795e0,).	
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(%,4+7).-(%+	8 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(74).9 +(.% 0	635
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb()9,- 8	7%4%9%%.
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(49),,4-30378 052
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(%7863	
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(e0
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(%72+e%e,49 7 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(	55,.
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(263,%	1-74)e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb( 64110, (1 
Parsed as rgb(150,3,255): rgb( 59%3   725)
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(6e)-2- 180
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(,	.579e(( 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb()(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(74	e)e6e)-29e7
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(221%  (e(77
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(--4	e1%- )	8378 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(e4(01.e	
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb()6((9((.9 3-6438
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(),15%-+)(,).6
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(3	(-085932
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb((
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb()
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(e0944 6)
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb( 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(e5
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(3%+)2%3%
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(6e5%.3	e+ e1.69e16
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(1		5.56,+8 5e%,3
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(	)ee.%)
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(-%0%-
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(231293 16%
Parsed as rgb(0,0,37): rgb(+-2+37)888580+9	(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(18	805
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(0
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(63-.693+
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(80,2-(.4		,
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(5%(59%2	5
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(5e3%,9
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(%7e-++59e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(963 (8)%73)	ee(-
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(71+1+()3382 3
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(.27+
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb((
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(88
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(9.43,(6.87 -
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(e)e. 5997
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(5 9-557
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb((65-%-0,e217
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb( 2
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb((e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(	8(+((e%52, .73
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(,0(e 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb()59)62159544,8,e3e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(0857-.5
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(6,71(5	.-%.
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(2	5e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(1	2-7	+-e90
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(9 )-7
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(8	759 6,
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(2+65e 59%0(8)5
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb( 0712
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(	97,	33)-6,+%0)1e(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb( .3.6 	,-57%7 e+6
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(1)7 +%+8 	)-	 -.63e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(6786).(8912%
Parsed as rgb(0,8,0): rgb(-	%8-,)886+- 9e))e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb( ,	5732 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(18)(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(3,+13%(711%9%	%0 -
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(6e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(-e(69
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(,.19%e-2-(0)+9-2
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(1%,))(810-5  %37
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(35.-(%
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(362020+-8347081%78
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(-(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(%) 54+(006e3141%
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(	-e 1(9	 0-7
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(61-2	 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(7075)5+	67-(+)7-7-
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(	..
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(88-+3%21
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(3%)003+4
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(682).e19)0-4642
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(6.7
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb( 9	1 %6(,%0)e20%e-9
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(	(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(96,	2e0.1%2)5-
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(.8)7+3 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(8.78
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(2e8e2% 172e+0. -
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(e9
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(0.795e0,).	%%,
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(+7).
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb((%+	8 -74).9 +(.
Parsed as rgb(0,255,0): rgb( 0	635+)9,- 8	
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(%4%9%%.
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(49),,4-30378 052
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(%7863	
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(e0
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(%72+e%e,49 7 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(	55,.
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(263,%	1-74)e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb( 64110, (1 
Parsed as rgb(150,3,255): rgb( 59%3   725)
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(6e)-2- 180
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(,	.579e(( 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb()(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(74	e)e6e)-29e7
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(221%  (e(77
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(--4	e1%- )	8378 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(e4(01.e	
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb()6((9((.9 3-6438
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(),15%-+)(,).6
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(3	(-085932
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb((
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb()
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(e0944 6)
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb( 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(e5
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(3%+)2%3%
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(6e5%.3	e+ e1.69e16
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(1		5.56,+8 5e%,3
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(	)ee.%)
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(-%0%-
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(231293 16%
Parsed as rgb(0,0,37): rgb(+-2+37)888580+9	(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(18	805
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(0
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(63-.693+
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(80,2-(.4		,
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(5%(59%2	5
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(5e3%,9
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(%7e-++59e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(963 (8)%73)	ee(-
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(71+1+()3382 3
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(.27+
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb((
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(88
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(9.43,(6.87 -
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(e)e. 5997
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(5 9-557
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb((65-%-0,e217
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb( 2
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb((e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(	8(+((e%52, .73
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(,0(e 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb()59)62159544,8,e3e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(0857-.5
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(6,71(5	.-%.
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(2	5e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(1	2-7	+-e90
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(9 )-7
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(8	759 6,
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(2+65e 59%0(8)5
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb( 0712
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(	97,	33)-6,+%0)1e(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb( .3.6 	,-57%7 e+6
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(1)7 +%+8 	)-	 -.63e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(6786).(8912%
Parsed as rgb(0,8,0): rgb(-	%8-,)886+- 9e))e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb( ,	5732 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(18)(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(3,+13%(711%9%	%0 -
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(6e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(-e(69
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(,.19%e-2-(0)+9-2
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(1%,))(810-5  %37
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(35.-(%
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(362020+-8347081%78
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(-(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(%) 54+(006e3141%
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(	-e 1(9	 0-7
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(61-2	 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(7075)5+	67-(+)7-7-
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(	..
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(88-+3%21
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(3%)003+4
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(682).e19)0-4642
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(6.7
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb( 9	1 %6(,%0)e20%e-9
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(	(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(96,	2e0.1%2)5-
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(.8)7+3 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(8.78
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(2e8e2% 172e+0. -
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(e9
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(0.795e0,).	%%,
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(+7).
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb((%+	8 -74).9 +(.
Parsed as rgb(0,255,0): rgb( 0	635+)9,- 8	
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(%4%9%%.
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(49),,4-30378 052
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(%7863	
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(e0
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(%72+e%e,49 7 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(	55,.
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(263,%	1-74)e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb( 64110, (1 
Parsed as rgb(150,3,255): rgb( 59%3   725)
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(6e)-2- 180
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(,	.579e(( 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb()(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(74	e)e6e)-29e7
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(221%  (e(77
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(--4	e1%- )	8378 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(e4(01.e	
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb()6((9((.9 3-6438
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(),15%-+)(,).6
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(3	(-085932
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb((
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb()
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(e0944 6)
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb( 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(e5
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(3%+)2%3%
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(6e5%.3	e+ e1.69e16
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(1		5.56,+8 5e%,3
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(	)ee.%)
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(-%0%-
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(231293 16%
Parsed as rgb(0,0,37): rgb(+-2+37)888580+9	(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(18	805
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(0
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(63-.693+
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(80,2-(.4		,
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(5%(59%2	5
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(5e3%,9
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(%7e-++59e
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(963 (8)%73)	ee(-
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(71+1+()3382 3
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(.27+
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb(
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: 
Threw exception Error: SVG_INVALID_VALUE_ERR: DOM SVG Exception 1: rgb()
PASS successfullyParsed is true

TEST COMPLETE
